requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0
jsonschema==4.20.0

# CLI & API
click==8.1.7
flask==3.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Report Generation
reportlab==4.0.7
python-docx==0.8.11
jinja2==3.1.2
weasyprint==59.0

# Security & Cryptography
cryptography==41.0.7
pycryptodome==3.18.0

# Data Processing
pandas==2.1.3
numpy==1.24.3

# Web Requests
requests==2.31.0
aiohttp==3.9.1

# Database (optional)
sqlalchemy==2.0.23
sqlite3-python==1.0.0

# Logging & Monitoring
python-json-logger==2.0.7
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Performance
lxml==4.9.3
psutil==5.9.6

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Data Serialization
msgpack==1.0.7

# Dashboard (optional)
plotly==5.18.0
dash==2.14.1

# APK/IPA Analysis (optional)
androguard==4.1.0
# objection would be installed via pip or as system dependency
# mobsf would be Docker-based
# frida would be installed via pip: frida, frida-tools

# Optional extras for advanced features
# google-cloud-storage==2.14.0  # For cloud integration
# redis==5.0.1  # For caching
# celery==5.3.4  # For distributed tasks
